Release Snorlax 0.4
New CLI to run the suggestion engine for users
Some listeners to clean up suggested friends
Updates to available games
Delete Suggestion When Added To Friend Table
Created a new listener on attach.friend.post to call deleteSuggestions
When a user sends a friend request or receive a friend request, that suggestion is deleted.
When a user accepts a friend request and that suggestion is still in the table, it is deleted.
Added unit test for the listener.
Added integration test to see that the suggestion is actually being deleted.
Added an eventIdentifier property to friendservicedelegator
Modifying the test to attach the listener to an event manager and let the event manager trigger the event
Suggest Cron Job which injects suggest Jobs for users in the database
Adding the console usage messages and refactoring the commands